Pair cmd/server/app test files with split source files
